requirements.txt
comtypes==1.1.9
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pypiwin32==223
pytesseract==0.3.7
pyttsx3==2.90
pywin32==300
